--- src/main/java/com/solvd/banksystemparser/print/Printable.java
@@ -1,8 +1,0 @@
-package com.solvd.banksystemparser.print;
-
-@FunctionalInterface
-public interface Printable {
-
-    void print();
-
-}
